Get rid of _extra.c files in core/unit.
Entire-Checkpoint: 01M23Z4JREFV4T7MY83E2D9EWC
