+++ core/unit/ctest_array_ops.c
@@ -1142,6 +1142,118 @@ void test_array_min_range_ho() {
   test_array_min_range(false);
 }
 
+void
+test_array_new_meta()
+{
+  struct gkyl_array *a = gkyl_array_new(GKYL_DOUBLE, 3, 10);
+  TEST_CHECK( a->type == GKYL_DOUBLE );
+  TEST_CHECK( a->ncomp == 3 );
+  TEST_CHECK( a->size == 10 );
+  TEST_CHECK( gkyl_array_is_cu_dev(a) == false );
+  gkyl_array_release(a);
+}
+
+void
+test_array_clear_basic()
+{
+  struct gkyl_array *a = gkyl_array_new(GKYL_DOUBLE, 1, 50);
+  gkyl_array_clear(a, 3.5);
+  double *d = a->data;
+  for (unsigned i=0; i<a->size*a->ncomp; ++i)
+    TEST_CHECK( d[i] == 3.5 );
+  gkyl_array_release(a);
+}
+
+void
+test_array_scale_basic()
+{
+  struct gkyl_array *a = gkyl_array_new(GKYL_DOUBLE, 1, 20);
+  gkyl_array_clear(a, 2.0);
+  gkyl_array_scale(a, 4.0);
+  double *d = a->data;
+  for (unsigned i=0; i<a->size; ++i)
+    TEST_CHECK( gkyl_compare_double(d[i], 8.0, 1e-14) );
+  gkyl_array_release(a);
+}
+
+void
+test_array_accumulate_basic()
+{
+  // out = out + a*inp
+  struct gkyl_array *out = gkyl_array_new(GKYL_DOUBLE, 1, 20);
+  struct gkyl_array *inp = gkyl_array_new(GKYL_DOUBLE, 1, 20);
+  gkyl_array_clear(out, 1.0);
+  gkyl_array_clear(inp, 2.0);
+  gkyl_array_accumulate(out, 3.0, inp);
+  double *d = out->data;
+  for (unsigned i=0; i<out->size; ++i)
+    TEST_CHECK( gkyl_compare_double(d[i], 1.0 + 3.0*2.0, 1e-14) );
+  gkyl_array_release(out); gkyl_array_release(inp);
+}
+
+void
+test_array_set_basic()
+{
+  // out = a*inp
+  struct gkyl_array *out = gkyl_array_new(GKYL_DOUBLE, 1, 20);
+  struct gkyl_array *inp = gkyl_array_new(GKYL_DOUBLE, 1, 20);
+  gkyl_array_clear(out, 99.0);
+  gkyl_array_clear(inp, 5.0);
+  gkyl_array_set(out, 2.0, inp);
+  double *d = out->data;
+  for (unsigned i=0; i<out->size; ++i)
+    TEST_CHECK( gkyl_compare_double(d[i], 10.0, 1e-14) );
+  gkyl_array_release(out); gkyl_array_release(inp);
+}
+
+void
+test_array_shiftc_basic()
+{
+  // shift component k by a (out[k] += a)
+  struct gkyl_array *a = gkyl_array_new(GKYL_DOUBLE, 3, 10);
+  gkyl_array_clear(a, 0.0);
+  gkyl_array_shiftc(a, 7.0, 1); // shift component 1
+  for (unsigned i=0; i<a->size; ++i) {
+    double *row = gkyl_array_fetch(a, i);
+    TEST_CHECK( row[0] == 0.0 );
+    TEST_CHECK( gkyl_compare_double(row[1], 7.0, 1e-14) );
+    TEST_CHECK( row[2] == 0.0 );
+  }
+  gkyl_array_release(a);
+}
+
+void
+test_array_copy_basic()
+{
+  struct gkyl_array *a = gkyl_array_new(GKYL_DOUBLE, 2, 15);
+  struct gkyl_array *b = gkyl_array_new(GKYL_DOUBLE, 2, 15);
+  double *ad = a->data;
+  for (unsigned i=0; i<a->size*a->ncomp; ++i) ad[i] = 0.5*i;
+  gkyl_array_copy(b, a);
+  double *bd = b->data;
+  for (unsigned i=0; i<b->size*b->ncomp; ++i)
+    TEST_CHECK( bd[i] == 0.5*i );
+  gkyl_array_release(a); gkyl_array_release(b);
+}
+
+void
+test_array_fetch_multicomp()
+{
+  // Write per-component values via fetch and read back.
+  struct gkyl_array *a = gkyl_array_new(GKYL_DOUBLE, 4, 8);
+  for (unsigned i=0; i<a->size; ++i) {
+    double *row = gkyl_array_fetch(a, i);
+    for (unsigned c=0; c<a->ncomp; ++c)
+      row[c] = 100.0*i + c;
+  }
+  for (unsigned i=0; i<a->size; ++i) {
+    const double *row = gkyl_array_cfetch(a, i);
+    for (unsigned c=0; c<a->ncomp; ++c)
+      TEST_CHECK( row[c] == 100.0*i + c );
+  }
+  gkyl_array_release(a);
+}
+
 // Cuda specific tests
 #ifdef GKYL_HAVE_CUDA
 
@@ -2184,6 +2296,14 @@ TEST_LIST = {
   { "array_copy_range", test_array_copy_range},
   { "array_copy_split", test_array_copy_split },
   { "array_copy_range_to_range_diff_range_dim", test_array_copy_range_to_range_diff_range_dim_ho},
+  { "array_new_meta", test_array_new_meta },
+  { "array_clear_basic", test_array_clear_basic },
+  { "array_scale_basic", test_array_scale_basic },
+  { "array_accumulate_basic", test_array_accumulate_basic },
+  { "array_set_basic", test_array_set_basic },
+  { "array_shiftc_basic", test_array_shiftc_basic },
+  { "array_copy_basic", test_array_copy_basic },
+  { "array_fetch_multicomp", test_array_fetch_multicomp },
 #ifdef GKYL_HAVE_CUDA
   { "cu_array_clear", test_cu_array_clear},
   { "cu_array_clear_range", test_cu_array_clear_range},
